[type: refactor] refactor shenyu sdk client
