+++ shenyu-sdk/shenyu-sdk-httpclient/pom.xml
@@ -19,20 +19,12 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>shenyu-sdk</artifactId>
         <groupId>org.apache.shenyu</groupId>
+        <artifactId>shenyu-sdk</artifactId>
         <version>2.5.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-
-
     <artifactId>shenyu-sdk-httpclient</artifactId>
-
-    <properties>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
 
     <dependencies>
         <dependency>
@@ -49,5 +41,4 @@
             <artifactId>httpcore</artifactId>
         </dependency>
     </dependencies>
-
 </project>
