+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-sdk/src/main/resources/META-INF/spring.factories
@@ -16,4 +16,4 @@
 #
 
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.apache.shenyu.springboot.starter.sdk.ShenyuClientAutoConfiguration
+org.apache.shenyu.springboot.starter.sdk.ShenyuSdkAutoConfiguration
